prompt adjusts
