+++ generators/server/command.js
@@ -1,9 +1,21 @@
+import { command as serverCommand } from 'generator-jhipster/generators/server';
+
+const { applicationType, ...remainingConfigs } = serverCommand.configs;
+
 /**
  * @type {import('generator-jhipster').JHipsterCommandDefinition}
  */
 const command = {
-  options: {},
-  configs: {},
+  ...serverCommand,
+  configs: {
+    // Gateway is not supported
+    applicationType: {
+      ...applicationType,
+      choices: applicationType.choices.filter(({ value }) => value !== 'gateway'),
+    },
+    ...remainingConfigs,
+  },
+  override: true,
   compose: ['jhipster-micronaut:micronaut'],
 };
