+++ generators/server/generator.js
@@ -5,6 +5,16 @@ export default class extends BaseApplicationGenerator {
     super(args, opts, { ...features, sbsBlueprint: true });
   }
 
+  get [BaseApplicationGenerator.INITIALIZING]() {
+    return this.asInitializingTaskGroup({
+      async initializingTemplateTask() {
+        // Disables some prompts by making it configured, it will be shown using `--ask-answered` option.
+        this.jhipsterConfig.reactive = false;
+        this.jhipsterConfig.databaseType = 'sql';
+      },
+    });
+  }
+
   get [BaseApplicationGenerator.CONFIGURING]() {
     return this.asPreparingTaskGroup({
       async preparingMicronaut() {
